+++ Source/Models/TransactionConsumptionParams.swift
@@ -17,10 +17,6 @@ public struct TransactionConsumptionParams {
     public let amount: BigInt?
     /// The address to use for the consumption
     public let address: String?
-    /// The id of the token to use for the request
-    /// In the case of a type "send", this will be the token that the consumer will receive
-    /// In the case of a type "receive" this will be the token that the consumer will send
-    public let tokenId: String?
     /// The idempotency token to use for the consumption
     public let idempotencyToken: String
     /// An id that can uniquely identify a transaction. Typically an order id from a provider.
@@ -36,16 +32,12 @@ public struct TransactionConsumptionParams {
     /// - Parameters:
     ///   - transactionRequest: The transaction request to consume
     ///   - address: The address to use for the consumption
-    ///   - tokenId: The id of the token to use for the request
-    ///              In the case of a type "send", this will be the token that the consumer will receive
-    ///              In the case of a type "receive" this will be the token that the consumer will send
     ///   - amount: The amount of token to transfer (down to subunit to unit)
     ///   - idempotencyToken: The idempotency token to use for the consumption
     ///   - correlationId: An id that can uniquely identify a transaction. Typically an order id from a provider.
     ///   - metadata: Additional metadata for the consumption
     public init?(transactionRequest: TransactionRequest,
                  address: String? = nil,
-                 tokenId: String? = nil,
                  amount: BigInt?,
                  idempotencyToken: String,
                  correlationId: String? = nil,
@@ -55,7 +47,6 @@ public struct TransactionConsumptionParams {
         self.formattedTransactionRequestId = transactionRequest.formattedId
         self.amount = amount == transactionRequest.amount ? nil : amount
         self.address = address
-        self.tokenId = tokenId
         self.idempotencyToken = idempotencyToken
         self.correlationId = correlationId
         self.metadata = metadata
@@ -70,7 +61,6 @@ extension TransactionConsumptionParams: APIParameters {
         case formattedTransactionRequestId = "formatted_transaction_request_id"
         case amount
         case address
-        case tokenId = "token_id"
         case metadata
         case encryptedMetadata = "encrypted_metadata"
         case correlationId = "correlation_id"
@@ -82,7 +72,6 @@ extension TransactionConsumptionParams: APIParameters {
         try container.encode(formattedTransactionRequestId, forKey: .formattedTransactionRequestId)
         try container.encode(amount, forKey: .amount)
         try container.encode(address, forKey: .address)
-        try container.encode(tokenId, forKey: .tokenId)
         try container.encode(metadata, forKey: .metadata)
         try container.encode(encryptedMetadata, forKey: .encryptedMetadata)
         try container.encode(correlationId, forKey: .correlationId)