+++ CHANGELOG.md
@@ -5,10 +5,23 @@ All notable changes to this project will be documented in this file.
 #### 0.x Betas
 - `0.10.x` Betas
   - [0.10.0](#0100)
+  - [0.10.1](#0101)
 
 
 - `0.9.x` Betas
   - [0.9.11](#0911)
+
+---
+## [0.10.1](https://github.com/omisego/ios-sdk/releases/tag/0.10.1)
+Released on 2018-6-27. All issues associated with this milestone can be found using this [filter](https://github.com/omisego/ios-sdk/issues?utf8=%E2%9C%93&q=milestone%3A0.10.1).
+
+#### Updated
+- decoding strategy for BigInt.
+  - Updated by [Mederic](https://github.com/mederic-p) in Pull Request [#63](https://github.com/omisego/ios-sdk/pull/63)
+
+#### Added
+- Support for server exchange.
+  - Added by [Mederic](https://github.com/mederic-p) in Pull Request [#64](https://github.com/omisego/ios-sdk/pull/64).
 
 ---
 ## [0.10.0](https://github.com/omisego/ios-sdk/releases/tag/0.10.0)